luongggg

+++ controller/ProfileController.php
@@ -3,10 +3,11 @@
 if (isset($_GET['act'])){
     switch ($_GET['act']) {
             case 'detail':
-
+                include_once 'model/user.php';
                 include_once 'model/checkout.php';
                 $getBillByClientId=getBillByClientId();
 
+                $clientInfor = getClientById($_SESSION['user']['client_id']);
                 include_once 'controller/CartController.php'; 
                 if (isset($_SESSION['cart']) && !empty($_SESSION['cart'])) {
                     function countCartItems1() {
@@ -76,35 +77,39 @@ if (isset($_GET['act'])){
                     include_once 'views/profile.php';
                     include_once 'views/layout/footer.php';
                 } else {
-                    if (isset($_GET['edit']) && isset($_GET['id'])&&isset($_POST['update'])) {
-                        $name = $_POST['name'];
-                $id = $_POST['id'];
-                $clientInfor = getClientById($_POST['id']);
-                $username = $_POST['username'];
-                $password = $_POST['password'];
-                $role = $_POST['role'];
-                $email = $_POST['email'];
-                $birthday = $_POST['birthday'];
-                $address = $_POST['address'];
-                $number = $_POST['number'];
-                $fileName = null;
-                if (isset($_FILES['image']) && $_FILES['image']['name'] != '') {
-                    $filePath = 'upload/AnhNhanBan/client/';
-            
-                    $fileName = date("Y_m_d_H_i_s") . $_FILES['image']['name'];
-                    // Xóa file cũ nếu có. Gợi ý if (file_exist())
-                    move_uploaded_file($_FILES['image']['tmp_name'], $filePath . $fileName);
-                // Kiểm tra nếu file đã tồn tại (tùy chọn: có thể thêm kiểm tra để xóa file cũ nếu cần)
-                unlink($filePath . $clientInfor['image']);
+                    if ($_GET['url'] == 'profile' && $_GET['act'] == 'edit') {
+                        if (isset($_POST['update'])) {
+                            $name = $_POST['name'];
+                            $id = $_POST['id'];
+                            $clientInfor = getClientById($_POST['id']);
+                            $password = $_POST['password'];
+                            $email = $_POST['email'];
+                            $address = $_POST['address'];
+                            $number = $_POST['number'];
+                            $birthday = $_POST['birthday'];
+                            $fileName = null;
+                            if (isset($_FILES['image']) && $_FILES['image']['name'] != '') {
+                                $filePath = 'admin/upload/AnhNhanBan/client/';
+                        
+                                $fileName = date("Y_m_d_H_i_s") . $_FILES['image']['name'];
+                                // Xóa file cũ nếu có. Gợi ý if (file_exist())
+                                move_uploaded_file($_FILES['image']['tmp_name'], $filePath . $fileName);
+                            // Kiểm tra nếu file đã tồn tại (tùy chọn: có thể thêm kiểm tra để xóa file cũ nếu cần)
+                            if (isset($clientInfor['image']) && file_exists($filePath . $clientInfor['image'])) {
+                                unlink($filePath . $clientInfor['image']);
+                            }
+                            }
 
-                }
-                editClient($id,$username, $password, $name, $email,$role ,$fileName,$address,$number,$birthday);
-                $script = "<script> 
-                alert('Sửa thông tin người dùng thành công!');
-                window.location = 'http://localhost/duan1_nhom4/index.php?url=profile&act=detail';
-                </script>";
-                echo $script;
+                            editClient($id, $password, $name, $email,$fileName,$address,$number,$birthday);
+                            $script = "<script> 
+                            alert('Sửa thông tin người dùng thành công!');
+                            window.location = 'index.php?url=profile&act=detail';
+                            </script>";
+                            echo $script;
+                        }
                     }
+                    
+
                     include_once 'controller/CartController.php'; 
                     if (isset($_SESSION['cart']) && !empty($_SESSION['cart'])) {
                         function countCartItems1() {
